requirements-dev.txt
# ============================================================================
# NjordScan Development Dependencies
# Additional packages needed for development, testing, and documentation
# ============================================================================

# Testing Framework
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
pytest-cov>=4.1.0,<5.0.0
pytest-xdist>=3.3.0,<4.0.0
pytest-mock>=3.11.0,<4.0.0
pytest-benchmark>=4.0.0,<5.0.0
pytest-timeout>=2.1.0,<3.0.0

# Code Formatting and Linting
black>=23.7.0,<24.0.0
flake8>=6.0.0,<7.0.0
isort>=5.12.0,<6.0.0
mypy>=1.5.0,<2.0.0
bandit>=1.7.0,<2.0.0
pylint>=2.17.0,<3.0.0
pycodestyle>=2.11.0,<3.0.0
pydocstyle>=6.3.0,<7.0.0

# Security and Quality
safety>=3.0.0,<4.0.0
semgrep>=1.40.0,<2.0.0
pre-commit>=3.3.0,<4.0.0
commitizen>=3.6.0,<4.0.0

# Coverage and Reporting
coverage>=7.3.0,<8.0.0
coverage-badge>=1.1.0,<2.0.0
codecov>=2.1.0,<3.0.0

# Documentation
mkdocs>=1.5.0,<2.0.0
mkdocs-material>=9.2.0,<10.0.0
mkdocs-mermaid2-plugin>=1.1.0,<2.0.0
mkdocs-git-revision-date-localized-plugin>=1.2.0,<2.0.0
mkdocs-minify-plugin>=0.7.0,<1.0.0
mkdocs-awesome-pages-plugin>=2.9.0,<3.0.0
mkdocs-include-markdown-plugin>=6.0.0,<7.0.0

# Type Checking and Stubs
types-requests>=2.31.0,<3.0.0
types-PyYAML>=6.0.0,<7.0.0
types-toml>=0.10.0,<1.0.0
types-beautifulsoup4>=4.12.0,<5.0.0
types-Pillow>=10.0.0,<11.0.0

# Development Tools
ipython>=8.14.0,<9.0.0
ipdb>=0.13.0,<1.0.0
pdbpp>=0.10.0,<1.0.0
rich-traceback>=0.1.0,<1.0.0

# Build and Packaging
build>=0.10.0,<1.0.0
twine>=4.0.0,<5.0.0
wheel>=0.41.0,<1.0.0
setuptools>=68.0.0,<69.0.0
setuptools-scm>=7.1.0,<8.0.0

# Performance Profiling
py-spy>=0.3.0,<1.0.0
memory-profiler>=0.61.0,<1.0.0
line-profiler>=4.1.0,<5.0.0
scalene>=1.5.0,<2.0.0

# Mock and Test Data
factory-boy>=3.3.0,<4.0.0
faker>=19.3.0,<20.0.0
responses>=0.23.0,<1.0.0
httpretty>=1.1.0,<2.0.0

# Jupyter and Notebooks (for analysis and demos)
jupyter>=1.0.0,<2.0.0
jupyterlab>=4.0.0,<5.0.0
notebook>=7.0.0,<8.0.0
nbconvert>=7.7.0,<8.0.0

# Additional Development Utilities
watchdog>=3.0.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0
click-completion>=0.5.0,<1.0.0
python-json-logger>=2.0.0,<3.0.0
structlog>=23.1.0,<24.0.0

# Database Tools for Development
alembic>=1.11.0,<2.0.0
sqlite-utils>=3.34.0,<4.0.0

# Network Testing
httpx>=0.24.0,<1.0.0
aioresponses>=0.7.0,<1.0.0
pytest-httpserver>=1.0.0,<2.0.0

# Benchmarking and Performance
asv>=0.6.0,<1.0.0
pytest-benchmark>=4.0.0,<5.0.0

# Git and Version Control Tools
gitpython>=3.1.0,<4.0.0
pre-commit-hooks>=4.4.0,<5.0.0

# Documentation Building Tools
sphinx>=7.1.0,<8.0.0
sphinx-rtd-theme>=1.3.0,<2.0.0
sphinx-autodoc-typehints>=1.24.0,<2.0.0

# API Documentation
pdoc>=14.0.0,<15.0.0
pydoc-markdown>=4.8.0,<5.0.0

# Configuration Management
dynaconf>=3.2.0,<4.0.0
python-decouple>=3.8,<4.0.0

# Async Testing Utilities
pytest-asyncio-cooperative>=0.21.0,<1.0.0
asynctest>=0.13.0,<1.0.0

# Load Testing (for performance testing)
locust>=2.16.0,<3.0.0

# Security Testing Tools
bandit[toml]>=1.7.0,<2.0.0
pip-audit>=2.6.0,<3.0.0

# Code Quality Metrics
radon>=6.0.0,<7.0.0
xenon>=0.9.0,<1.0.0
vulture>=2.9.0,<3.0.0

# Dependency Management
pip-tools>=7.1.0,<8.0.0
pipdeptree>=2.13.0,<3.0.0

# Container and Docker Support (for testing)
docker>=6.1.0,<7.0.0
testcontainers>=3.7.0,<4.0.0

# Continuous Integration Tools
tox>=4.10.0,<5.0.0
nox>=2023.4.22,<2024.0.0

# Additional Linting and Formatting
autopep8>=2.0.0,<3.0.0
yapf>=0.40.0,<1.0.0
pyupgrade>=3.10.0,<4.0.0

# Import Sorting and Organization
autoflake>=2.2.0,<3.0.0
unify>=0.5,<1.0.0

# Documentation Linting
doc8>=1.1.0,<2.0.0
rstcheck>=6.1.0,<7.0.0

# License Checking
licensecheck>=2023.1.1,<2024.0.0

# Changelog Generation
auto-changelog>=0.6.0,<1.0.0
towncrier>=23.6.0,<24.0.0

# Release Management
bump2version>=1.0.0,<2.0.0
semantic-version>=2.10.0,<3.0.0

# Environment Management
virtualenv>=20.24.0,<21.0.0
pipenv>=2023.7.23,<2024.0.0

# Shell Integration and Completion
argcomplete>=3.1.0,<4.0.0
shellingham>=1.5.0,<2.0.0
